Refactor of changes request
