+++ e2e/features/board.feature
@@ -3,30 +3,30 @@ Feature: Create Dashboard in Trello
   As a User
   I want to create dashboards in Trello
 
-  Scenario: Create a Dashboard in Trello
+  Scenario: Create a board in Trello
     Given I login "https://trello.com/login?returnUrl=%2F"
-    When I create a new Dashboard with:
-      | Title       |   wdio2Board  |
-      | Privacy     |   Public      |
-      | Background  |   orange      |
+    When I create a new Board with:
+      | Title       |   newBoard     |
+      | Privacy     |   Public       |
+      | Background  |   orange       |
     Then I expect my board created
 
-  Scenario: Create a public Dashboard in Trello
+  Scenario: Create a board in Trello
     Given I login "https://trello.com/login?returnUrl=%2F"
-    When I create a new Dashboard with:
+    When I create a new Board with:
       | Title       |   NameBoard  |
     Then I expect my board created
 
-  Scenario: Create a public Dashboard in Trello
+  Scenario: Create a public board in Trello
     Given I login "https://trello.com/login?returnUrl=%2F"
-    When I create a new Dashboard with:
+    When I create a new Board with:
       | Title       |   publicBoard  |
       | Privacy     |   Public       |
     Then I expect my board created
 
-  Scenario: Create a public Dashboard in Trello
+  Scenario: Create a private board in Trello
     Given I login "https://trello.com/login?returnUrl=%2F"
-    When I create a new Dashboard with:
+    When I create a new Board with:
       | Title       |   privateBoard  |
       | Privacy     |   Private       |
     Then I expect my board created
